mate: add missing 'fi', make autologin match xfce, fix perms for homedir
Some fixes for the MATE spin. There was a duplicate 'sed' run on
liveinst.desktop, one of which was in an 'if' statement with no 'fi'; this
meant livesys never completed. Took out one of the runs and added the 'fi'
to make livesys complete. Also changed to the style from the Xfce
kickstart for LightDM autologin (seems sensible to have it consistent
between spins, and we know the Xfce method works) and added the stanza for
correcting ownership and SELinux context of /home/liveuser that is present
in all other desktop live spins.