slightly ugly hack to give us a working SoaS for beta: s/gdm/lightdm/
This quick hack should give us a working SoaS image for beta, I tested it
here. Use lightdm instead of gdm and adjust for lightdm autologin. Seems
to work, and it was utterly broken in Beta RC3, so what the hell, give it
a shot. For Final we should change to lightdm in comps, if switching
away from gdm is what we decide to do.