Use git version to get version name
Using git version is simpler as it handles oddball cases internally.
Since we have been using mostly lightweight tags I have set it to
search those in addition to annotated tags for the best match. The
output for commits that don't have a tag seems a bit nicer as well.

Conflicts:
	Makefile